Change deserialization type from `BigInt` to `usize` (#499)
* Change `BigInt` type to `usize`.

* Fix clippy warning.

+++ src/serde/deserialize_program.rs
@@ -72,8 +72,7 @@ pub struct Identifier {
 #[derive(Clone, Debug, Deserialize, PartialEq)]
 pub struct Member {
     pub cairo_type: Option<String>,
-    #[serde(deserialize_with = "bigint_from_number")]
-    pub offset: Option<BigInt>,
+    pub offset: Option<usize>,
 }
 
 fn bigint_from_number<'de, D>(deserializer: D) -> Result<Option<BigInt>, D::Error>
